Cadastrar despesas domésticas no cartão de crédito (#153)

+++ Tests/Mocks/Database/Tables/202112152009_ReminingBalance.cs
@@ -10,8 +10,8 @@ namespace Cashflow.Tests.Mocks.Database.Tables
             Create.Table("RemainingBalance")
                 .WithColumn("Id").AsInt32().PrimaryKey().Identity()
                 .WithColumn("Value").AsDecimal(10, 2).NotNullable()
-                .WithColumn("Month").AsInt16().NotNullable()
-                .WithColumn("Year").AsInt16().NotNullable();
+                .WithColumn("Month").AsInt32().NotNullable()
+                .WithColumn("Year").AsInt32().NotNullable();
 
             Execute.Sql("ALTER TABLE RemainingBalance ADD COLUMN UserId INTEGER REFERENCES User(Id)");
         }
